refactor: decompose useWebEngine
